==> notebooks/2_TrainTestSplit.ipynb <==
# 2 Train Dev Test split

I split the train set further into my own labelled test set, of same size as develop set, as I like to have an extra set to evaluate the performance of my model beyond the dev set, which just selects the best parameters

- from now on: the original test set will be referred to as future set; the new test set (split off from train), which has labels, will be referred to as test set

import json
from sklearn.model_selection import train_test_split

with open ('../data/raw/train-claims.json') as f:
    train_claims = json.load(f)

with open ('../data/raw/dev-claims.json') as f:
    dev_claims = json.load(f)

# test size 154 because that is dev set size

train_data_index, test_data_index = train_test_split(list(train_claims), test_size=154, random_state = 19260817)

train_data = {index:train_claims[index] for index in train_data_index}
test_data = {index:train_claims[index] for index in test_data_index}

# 0R

# label them as '2' because they are not the original versions
with open('../data/curated/train_claims2.json', 'w') as f:
    json.dump(train_data, f)

with open('../data/curated/test_claims2.json', 'w') as f:
    json.dump(test_data, f)

from collections import defaultdict as dd

train_tally = dd(int)
for doc in train_data:

    train_tally[train_data[doc]['claim_label']] += 1

print(train_tally)


dev_tally = dd(int)
for doc in dev_claims:
    dev_tally[dev_claims[doc]['claim_label']] += 1

print(dev_tally)


test_tally = dd(int)
for doc in test_data:
    test_tally[test_data[doc]['claim_label']] += 1

print(test_tally)
