==> notebooks/3_1_Clustering.ipynb <==
# 3.1 Clustering: TFIDF + SVD; KNN and DOC2VEC; KNN

- tried to cluster to form a small cluster of evidence which contains all the seen evidences, and is size < 10000

import json
import nltk
from sklearn.feature_extraction.text import TfidfVectorizer
from nltk.stem import *
from nltk.corpus import stopwords
import string
import pickle

## Load in Data

with open ('../data/raw/train-claims.json') as f:
    train_claims = json.load(f)

with open ('../data/raw/dev-claims.json') as f:
    dev_claims = json.load(f)

with open ('../data/raw/evidence.json') as f:
    evidence = json.load(f)

evidence = list(evidence.items())
evidence.sort()

## Prepare Evidence

# ground truth evidences
definitely_related_evidence = set()
for claim in train_claims:
    for evidence in train_claims[claim]['evidences']:
        definitely_related_evidence.add(evidence)

definitely_related_evidence2 = set()
for claim in dev_claims:
    for evidence in dev_claims[claim]['evidences']:
        definitely_related_evidence2.add(evidence)

# 'train'
cluster_train = definitely_related_evidence.difference(definitely_related_evidence2)

# 'dev'
cluster_dev = definitely_related_evidence2

len(definitely_related_evidence)

len(definitely_related_evidence2)

len(cluster_train)

len(cluster_dev)

# all seen evidences
cluster_full = definitely_related_evidence.union(definitely_related_evidence2)

len(cluster_full)

---

## Preprocessing: With punctuation removal, stopword removal, stemming/lemmatization, 

nltk.download('stopwords')

ALL_PUNCT = '''!()-[]{};:'"\,<>./?@#$%^&*_~'''
stemmer = PorterStemmer()
STOP_WORDS = set(stopwords.words('english'))

def all_alpha(word):
    for char in word:
        if char not in string.ascii_lowercase:
            return False
    return True

def preprocess(text):
    text = text.lower()

    no_punct_text = str()
    for char in text:
        if char not in ALL_PUNCT:
            no_punct_text += char

    word_list = no_punct_text.split()

    no_stop_words_word_list = [word for word in word_list if word not in STOP_WORDS]

    stemmed_word_list = [stemmer.stem(word) for word in no_stop_words_word_list]

    output = ' '.join(stemmed_word_list)
    
    return output

EVIDENCE_ID_MAP = {x[1][0]:x[0] for x in enumerate(evidence)}

preprocessed_evidence = [preprocess(evidence[i][1]) for i in range(len(evidence))]

### TFIDF

vectorizer = TfidfVectorizer()
X = vectorizer.fit_transform(preprocessed_evidence)

with open('../data/curated/processed_tfidf_nonalpha.pickle', 'wb') as f:
    pickle.dump(X, f)

with open('../data/curated/processed_tfidf_nonalpha.pickle', 'rb') as f:
    X = pickle.load(f)

### SVD

from sklearn.decomposition import TruncatedSVD

svd = TruncatedSVD(5)

X_svd = svd.fit_transform(X)

X_svd

### Normalise

from sklearn import preprocessing

stdardiser = preprocessing.Normalizer()

X_svd_std = stdardiser.fit_transform(X_svd)

## Clustering using KMeans

from sklearn.cluster import KMeans
K = 2
cluster = KMeans(n_clusters = K, random_state = 19260817, n_init = 'auto').fit(X_svd)

labels = []
for evid in cluster_full:

    x = [X_svd[EVIDENCE_ID_MAP[evid]]]

    labels.append(cluster.predict(x)[0])


for i in set(cluster.labels_):
    print(i, labels.count(i))

all_labels = cluster.predict(X_svd)

# failed to get a small cluster of around 10000 that has all of the seen evidences
for i in set(cluster.labels_):
    print(i, list(all_labels).count(i))

---

# Doc2Vec

!pip install gensim

from gensim.test.utils import common_texts
from gensim.models.doc2vec import Doc2Vec, TaggedDocument

common_texts

evidence_tokenised = [evidence[i][1].split() for i in range(len(evidence))]

documents = [TaggedDocument(doc, [i]) for i, doc in enumerate(evidence_tokenised)]
model = Doc2Vec(documents, vector_size=2, window=2, min_count=1, workers=4)

X = list()

for i in range(len(evidence_tokenised)):
    X.append(model.infer_vector(evidence_tokenised[i]))

X

## K Means Clustering

from sklearn.cluster import KMeans
K = 2
cluster = KMeans(n_clusters = K, random_state = 19260817, n_init = 'auto').fit(X)

labels = []
for evid in cluster_full:

    x = [X[EVIDENCE_ID_MAP[evid]]]

    labels.append(cluster.predict(x)[0])


for i in set(cluster.labels_):
    print(i, labels.count(i))

all_labels = cluster.predict(X)

for i in set(cluster.labels_):
    print(i, list(all_labels).count(i))

# once again failed to form a small cluster that contains majority of the evidences
