==> notebooks/4_3_Retrieval_SBert.ipynb <==
# 4.3 Retrieval SBERT

from google.colab import drive
drive.mount('/content/drive')

## Read in Data

import json
import numpy as np
import gc
import pickle

with open ('./drive/My Drive/LAB/COMP90042 A3/data/curated/train_claims2.json') as f:
    train_claims = json.load(f)

with open ('./drive/My Drive/LAB/COMP90042 A3/data/raw/dev-claims.json') as f:
    dev_claims = json.load(f)

with open ('./drive/My Drive/LAB/COMP90042 A3/data/curated/test_claims2.json') as f:
    test_claims = json.load(f)

with open ('./drive/My Drive/LAB/COMP90042 A3/data/raw/test-claims-unlabelled.json') as f:
    future_claims = json.load(f)

with open ('./drive/My Drive/LAB/COMP90042 A3/data/raw/evidence.json') as f:
    evidence = json.load(f)

import random
random.seed(19260817)

## Reduce claims

evid_id_list = [evid_id for evid_id in evidence]

### Create dataset used to train retriever

NEG_SAMPLE_FACTOR = 1

training_data = []

for id in train_claims:

  claim_text = train_claims[id]['claim_text']

  n_evid = len(train_claims[id]['evidences'])

  for evid_id in train_claims[id]['evidences']:
    evid_text = evidence[evid_id]

    training_data.append((((claim_text, evid_text)), 1))
  
  #negative sampling
  neg_sampled = 0
  while neg_sampled < n_evid * NEG_SAMPLE_FACTOR:
    sampled_neg_evid_id = random.choice(evid_id_list)
    if sampled_neg_evid_id not in train_claims[id]['evidences']:
      neg_sampled += 1
      evid_text = evidence[sampled_neg_evid_id]

      training_data.append((((claim_text, evid_text)), 0))

dev_data = []

for id in dev_claims:

  claim_text = dev_claims[id]['claim_text']

  n_evid = len(dev_claims[id]['evidences'])

  for evid_id in dev_claims[id]['evidences']:
    evid_text = evidence[evid_id]

    dev_data.append((((claim_text, evid_text)), 1))
  
  #negative sampling
  neg_sampled = 0
  while neg_sampled < n_evid * NEG_SAMPLE_FACTOR:
    sampled_neg_evid_id = random.choice(evid_id_list)
    if sampled_neg_evid_id not in dev_claims[id]['evidences']:
      neg_sampled += 1
      evid_text = evidence[sampled_neg_evid_id]

      dev_data.append((((claim_text, evid_text)), 0))

test_data = []

for id in test_claims:

  claim_text = test_claims[id]['claim_text']

  n_evid = len(test_claims[id]['evidences'])

  for evid_id in test_claims[id]['evidences']:
    evid_text = evidence[evid_id]

    test_data.append((((claim_text, evid_text)), 1))
  
  #negative sampling
  neg_sampled = 0
  while neg_sampled < n_evid * NEG_SAMPLE_FACTOR:
    sampled_neg_evid_id = random.choice(evid_id_list)
    if sampled_neg_evid_id not in test_claims[id]['evidences']:
      neg_sampled += 1
      evid_text = evidence[sampled_neg_evid_id]

      test_data.append((((claim_text, evid_text)), 0))

!pip install torch torchvision transformers

## Build model, dataloader etc

from transformers import BertModel

from transformers import BertTokenizer

tokenizer = BertTokenizer.from_pretrained('bert-base-uncased')

import torch
from torch.utils.data import Dataset
from transformers import BertTokenizer
import pandas as pd

class Dataset():

    def __init__(self, data, maxlen):

        #Store the contents of the file in a pandas dataframe
        self.data = data

        #Initialize the BERT tokenizer
        self.tokenizer = BertTokenizer.from_pretrained('bert-base-uncased')

        self.maxlen = maxlen

    def __len__(self):
        return len(self.data)

    def __getitem__(self, index):

        #Selecting the sentence and label at the specified index in the data frame
        sentence1 = self.data[index][0][0]
        sentence2 = self.data[index][0][1]
        label = self.data[index][1]

        #Preprocessing the text to be suitable for BERT
        tokens1 = tokenizer.tokenize(sentence1)
        tokens2 = tokenizer.tokenize(sentence2)
        tokens1 = ['[CLS]'] + tokens1 + ['[SEP]']
        tokens2 = ['[CLS]'] + tokens2 + ['[SEP]']
        if len(tokens1) < self.maxlen:
            tokens1 = tokens1 + ['[PAD]' for _ in range(self.maxlen - len(tokens1))] #Padding sentences
        else:
            tokens1 = tokens1[:self.maxlen-1] + ['[SEP]'] #Prunning the list to be of specified max length

        if len(tokens2) < self.maxlen:
            tokens2 = tokens2 + ['[PAD]' for _ in range(self.maxlen - len(tokens2))] #Padding sentences
        else:
            tokens2 = tokens2[:self.maxlen-1] + ['[SEP]'] #Prunning the list to be of specified max length

        tokens_ids1 = self.tokenizer.convert_tokens_to_ids(tokens1) #Obtaining the indices of the tokens in the BERT Vocabulary
        tokens_ids_tensor1 = torch.tensor(tokens_ids1) #Converting the list to a pytorch tensor

        tokens_ids2 = self.tokenizer.convert_tokens_to_ids(tokens2) #Obtaining the indices of the tokens in the BERT Vocabulary
        tokens_ids_tensor2 = torch.tensor(tokens_ids2) #Converting the list to a pytorch tensor

        #Obtaining the attention mask i.e a tensor containing 1s for no padded tokens and 0s for padded ones
        attn_mask1 = (tokens_ids_tensor1 != 0).long()
        attn_mask2 = (tokens_ids_tensor2 != 0).long()



        return tokens_ids_tensor1, attn_mask1, tokens_ids_tensor2, attn_mask2,  label

from torch.utils.data import DataLoader

#Creating instances of training and development set
#maxlen sets the maximum length a sentence can have
#any sentence longer than this length is truncated to the maxlen size
train_set = Dataset(training_data, maxlen = 512)
dev_set = Dataset(dev_data, maxlen = 512)
test_set = Dataset(test_data, maxlen = 512)

#Creating intsances of training and development dataloaders
train_loader = DataLoader(train_set, batch_size = 4, shuffle = True, num_workers = 2)
dev_loader = DataLoader(dev_set, batch_size = 4, shuffle = True, num_workers = 2)
test_loader = DataLoader(test_set, batch_size = 4, shuffle = True, num_workers = 2)

import torch
import torch.nn as nn
from transformers import BertModel

class RelatednessClassifier(nn.Module):

    def __init__(self):
        super(RelatednessClassifier, self).__init__()
        #Instantiating BERT model object 
        self.bert_layer1 = BertModel.from_pretrained('bert-base-uncased')
        self.bert_layer2 = BertModel.from_pretrained('bert-base-uncased')

        
        #Classification layer
        #input dimension is 768 because [CLS] embedding has a dimension of 768
        #output dimension is 1 because we're working with a binary classification problem
        self.cls_layer = nn.Linear(1537, 1)

    def forward(self, seq1, attn_masks1, seq2, attn_masks2):
        '''
        Inputs:
            -seq : Tensor of shape [B, T] containing token ids of sequences
            -attn_masks : Tensor of shape [B, T] containing attention masks to be used to avoid contibution of PAD tokens
        '''
        batch_size = seq1.size(0)
        #Feeding the input to BERT model to obtain contextualized representations
        claim_outputs = self.bert_layer1(seq1, attention_mask = attn_masks1, return_dict=True)
        claim_cont_reps = claim_outputs.last_hidden_state

        #Obtaining the representation of [CLS] head (the first token)
        claim_cls_reps = claim_cont_reps[:, 0]

        #Feeding the input to BERT model to obtain contextualized representations
        evid_outputs = self.bert_layer2(seq2, attention_mask = attn_masks2, return_dict=True)
        evid_cont_reps = evid_outputs.last_hidden_state

        #Obtaining the representation of [CLS] head (the first token)
        evid_cls_reps = evid_cont_reps[:, 0]

        # Concatenate the two output tensors along the last dimension (i.e., the features dimension)
        concat_output = torch.cat((claim_cls_reps, evid_cls_reps), dim=-1)

        # Calculate the Euclidean distance between the two output tensors and flatten the result
        distances = []
        for i in range(batch_size):
            distance = torch.dist(claim_cls_reps[i], evid_cls_reps[i], p=2)
            distances.append(distance)
        distances = torch.flatten(torch.stack(distances)).unsqueeze(1)

        # Concatenate the flattened distance with the concatenated output tensor
        concat_output = torch.cat((concat_output, distances), dim=-1)

        #Feeding cls_rep to the classifier layer
        logits = self.cls_layer(concat_output)

        return logits
    

    def get_claim_embedding(self, seq1, attn_masks1):
        #Feeding the input to BERT model to obtain contextualized representations
        claim_outputs = self.bert_layer1(seq1, attention_mask = attn_masks1, return_dict=True)
        claim_cont_reps = claim_outputs.last_hidden_state

        #Obtaining the representation of [CLS] head (the first token)
        claim_cls_reps = claim_cont_reps[:, 0]

        return claim_cls_reps
    

    def get_evid_embedding(self, seq2, attn_masks2):
        #Feeding the input to BERT model to obtain contextualized representations
        evid_outputs = self.bert_layer2(seq2, attention_mask = attn_masks2, return_dict=True)
        evid_cont_reps = evid_outputs.last_hidden_state

        #Obtaining the representation of [CLS] head (the first token)
        evid_cls_reps = evid_cont_reps[:, 0]

        return evid_cls_reps
    

    def neural_layer(self, claim_cls_reps, evid_cls_reps, batch_size):

        # Concatenate the two output tensors along the last dimension (i.e., the features dimension)
        concat_output = torch.cat((claim_cls_reps, evid_cls_reps), dim=-1)

        # Calculate the Euclidean distance between the two output tensors and flatten the result
        distances = []
        for i in range(batch_size):
            distance = torch.dist(claim_cls_reps[i], evid_cls_reps[i], p=2)
            distances.append(distance)
        if batch_size == 1:
          distances = torch.flatten(torch.stack(distances))
        else:
          distances = torch.flatten(torch.stack(distances)).unsqueeze(1)

        # Concatenate the flattened distance with the concatenated output tensor
        concat_output = torch.cat((concat_output, distances), dim=-1)

        #Feeding cls_rep to the classifier layer
        logits = self.cls_layer(concat_output)

        return logits

torch.cuda.empty_cache()
gc.collect()

gpu = 0 #gpu ID

print("Creating the sentiment classifier, initialised with pretrained BERT-BASE parameters...")
net = RelatednessClassifier()
net.cuda(gpu) #Enable gpu support for the model
print("Done creating the sentiment classifier.")

### Setup Training

import torch.nn as nn
import torch.optim as optim

criterion = nn.BCEWithLogitsLoss()
opti = optim.Adam(net.parameters(), lr = 2e-5)

def get_accuracy_from_logits(logits, labels):
    probs = torch.sigmoid(logits.unsqueeze(-1))
    soft_probs = (probs > 0.5).long()
    acc = (soft_probs.squeeze() == labels).float().mean()
    return acc

def evaluate(net, criterion, dataloader, gpu):
    net.eval()

    mean_acc, mean_loss = 0, 0
    count = 0

    with torch.no_grad():
        for tokens_ids_tensor1, attn_mask1, tokens_ids_tensor2, attn_mask2, label in dataloader:
            tokens_ids_tensor1, attn_mask1, tokens_ids_tensor2, attn_mask2, label = tokens_ids_tensor1.cuda(gpu), attn_mask1.cuda(gpu), tokens_ids_tensor2.cuda(gpu), attn_mask2.cuda(gpu), label.cuda(gpu)
            logits = net(tokens_ids_tensor1, attn_mask1, tokens_ids_tensor2, attn_mask2)
            mean_loss += criterion(logits.squeeze(-1), label.float()).item()
            mean_acc += get_accuracy_from_logits(logits, label)
            count += 1

    return mean_acc / count, mean_loss / count

import time

def train(net, criterion, opti, train_loader, dev_loader, max_eps, gpu):

    best_acc = 0
    st = time.time()
    for ep in range(max_eps):
        
        net.train()
        for it, (tokens_ids_tensor1, attn_mask1, tokens_ids_tensor2, attn_mask2, label) in enumerate(train_loader):
            #Clear gradients
            opti.zero_grad()  
            #Converting these to cuda tensors
            tokens_ids_tensor1, attn_mask1, tokens_ids_tensor2, attn_mask2, label = tokens_ids_tensor1.cuda(gpu), attn_mask1.cuda(gpu), tokens_ids_tensor2.cuda(gpu), attn_mask2.cuda(gpu), label.cuda(gpu)

            #Obtaining the logits from the model
            logits = net(tokens_ids_tensor1, attn_mask1, tokens_ids_tensor2, attn_mask2)

            #Computing loss
            loss = criterion(logits.squeeze(-1), label.float())

            #Backpropagating the gradients
            loss.backward()

            #Optimization step
            opti.step()
              
            if it % 100 == 0:
                
                acc = get_accuracy_from_logits(logits, label)
                print("Iteration {} of epoch {} complete. Loss: {}; Accuracy: {}; Time taken (s): {}".format(it, ep, loss.item(), acc, (time.time()-st)))
                st = time.time()

        
        dev_acc, dev_loss = evaluate(net, criterion, dev_loader, gpu)
        print("Epoch {} complete! Development Accuracy: {}; Development Loss: {}".format(ep, dev_acc, dev_loss))
        if dev_acc > best_acc:
            print("Best development accuracy improved from {} to {}, saving model...".format(best_acc, dev_acc))
            best_acc = dev_acc
            torch.save(net.state_dict(), './drive/My Drive/LAB/COMP90042 A3/models/Retrievers/Retriever_SBert.dat')
            torch.save(net, './drive/My Drive/LAB/COMP90042 A3/models/Retrievers/Retriever_SBert.pt')

torch.cuda.empty_cache()
gc.collect()

num_epoch = 5

#fine-tune the model
train(net, criterion, opti, train_loader, dev_loader, num_epoch, gpu)

## Read In Model

net.load_state_dict(torch.load('./drive/My Drive/LAB/COMP90042 A3/models/Retrievers/Retriever_SBert.dat'))

net = torch.load('./drive/My Drive/LAB/COMP90042 A3/models/Retrievers/Retriever_SBert.pt')

net.eval()

## Make Predictions (to get Retrievals)

import torch
from torch.utils.data import Dataset
from transformers import BertTokenizer
import pandas as pd

class EmbeddingDataset():

    def __init__(self, data, maxlen):

        #Store the contents of the file in a pandas dataframe
        self.data = data

        #Initialize the BERT tokenizer
        self.tokenizer = BertTokenizer.from_pretrained('bert-base-uncased')

        self.maxlen = maxlen

    def __len__(self):
        return len(self.data)

    def __getitem__(self, index):

        #Selecting the sentence and label at the specified index in the data frame
        id = self.data[index][0]
        sentence1 = self.data[index][1]

        #Preprocessing the text to be suitable for BERT
        tokens1 = tokenizer.tokenize(sentence1)
        tokens1 = ['[CLS]'] + tokens1 + ['[SEP]']
        if len(tokens1) < self.maxlen:
            tokens1 = tokens1 + ['[PAD]' for _ in range(self.maxlen - len(tokens1))] #Padding sentences
        else:
            tokens1 = tokens1[:self.maxlen-1] + ['[SEP]'] #Prunning the list to be of specified max length

        tokens_ids1 = self.tokenizer.convert_tokens_to_ids(tokens1) #Obtaining the indices of the tokens in the BERT Vocabulary
        tokens_ids_tensor1 = torch.tensor(tokens_ids1) #Converting the list to a pytorch tensor

        #Obtaining the attention mask i.e a tensor containing 1s for no padded tokens and 0s for padded ones
        attn_mask1 = (tokens_ids_tensor1 != 0).long()



        return id, tokens_ids_tensor1, attn_mask1

dev_embedding_data = []

for id in dev_claims:

  claim_text = dev_claims[id]['claim_text']

  dev_embedding_data.append((id, claim_text))


test_embedding_data = []

for id in test_claims:

  claim_text = test_claims[id]['claim_text']

  test_embedding_data.append((id, claim_text))


future_embedding_data = []

for id in future_claims:

  claim_text = future_claims[id]['claim_text']

  future_embedding_data.append((id, claim_text))
  

evid_embedding_data = []
for evid in evidence:
  evid_text = evidence[evid]

  evid_embedding_data.append((evid, evid_text))

dev_embed_set = EmbeddingDataset(dev_embedding_data, maxlen = 512)
test_embed_set = EmbeddingDataset(test_embedding_data, maxlen = 512)
future_embed_set = EmbeddingDataset(future_embedding_data, maxlen = 512)
evid_embed_set = EmbeddingDataset(evid_embedding_data, maxlen = 512)

#Creating intsances of training and development dataloaders
dev_embed_loader = DataLoader(dev_embed_set, batch_size = 4, shuffle = False, num_workers = 2)
test_embed_loader = DataLoader(test_embed_set, batch_size = 4, shuffle = False, num_workers = 2)
future_embed_loader = DataLoader(future_embed_set, batch_size = 4, shuffle = False, num_workers = 2)
evid_embed_loader = DataLoader(evid_embed_set, batch_size = 4, shuffle = False, num_workers = 2)

import gc
torch.cuda.empty_cache()
gc.collect()

dev_embed = {}

for id, tokens_ids_tensor1, attn_mask1 in dev_embed_loader:
  tokens_ids_tensor1, attn_mask1 = tokens_ids_tensor1.cuda(gpu), attn_mask1.cuda(gpu)
  embed = net.get_claim_embedding(tokens_ids_tensor1, attn_mask1)

  embed = embed.tolist()
  for i in range(len(id)):
      dev_embed[id[i]] = embed[i]

with open('./drive/My Drive/LAB/COMP90042 A3/dev_embed.json', 'w') as f:
  json.dump(dev_embed, f)

test_embed = {}

for id, tokens_ids_tensor1, attn_mask1 in test_embed_loader:
  tokens_ids_tensor1, attn_mask1 = tokens_ids_tensor1.cuda(gpu), attn_mask1.cuda(gpu)
  embed = net.get_claim_embedding(tokens_ids_tensor1, attn_mask1)

  embed = embed.tolist()
  for i in range(len(id)):
      test_embed[id[i]] = embed[i]

with open('./drive/My Drive/LAB/COMP90042 A3/test_embed.json', 'w') as f:
  json.dump(test_embed, f)

future_embed = {}

for id, tokens_ids_tensor1, attn_mask1 in future_embed_loader:
  tokens_ids_tensor1, attn_mask1 = tokens_ids_tensor1.cuda(gpu), attn_mask1.cuda(gpu)
  embed = net.get_claim_embedding(tokens_ids_tensor1, attn_mask1)

  embed = embed.tolist()
  for i in range(len(id)):
      future_embed[id[i]] = embed[i]

with open('./drive/My Drive/LAB/COMP90042 A3/future_embed.json', 'w') as f:
  json.dump(future_embed, f)

torch.cuda.empty_cache()
gc.collect()

evid_embed = {}
j = 0
for id, tokens_ids_tensor1, attn_mask1 in evid_embed_loader:

  if j <= 172800:
    j+=1
    if j %6400 == 0:
      print('pass', j)

    continue

  tokens_ids_tensor1, attn_mask1 = tokens_ids_tensor1.cuda(gpu), attn_mask1.cuda(gpu)
  embed = net.get_evid_embedding(tokens_ids_tensor1, attn_mask1)
  embed = embed.tolist()

  torch.cuda.empty_cache()
  gc.collect()

  for i in range(len(id)):
      evid_embed[id[i]] = embed[i]
  
  j += 1
  if j %6400 == 0:
    print(j)
    with open(f'./drive/My Drive/LAB/COMP90042 A3/evid_embed{j}.json', 'w') as f:
      json.dump(evid_embed, f)
      evid_embed = {}

print(j)
with open(f'./drive/My Drive/LAB/COMP90042 A3/evid_embed{j}.json', 'w') as f:
  json.dump(evid_embed, f)
  evid_embed = {}

Prediction

with open('./drive/My Drive/LAB/COMP90042 A3/embeddings/other_embeddings/dev_embed.json', 'r') as f:
  dev_embed = json.load(f)

with open('./drive/My Drive/LAB/COMP90042 A3/embeddings/other_embeddings/test_embed.json', 'r') as f:
  test_embed = json.load(f)

with open('./drive/My Drive/LAB/COMP90042 A3/embeddings/other_embeddings/future_embed.json', 'r') as f:
  future_embed = json.load(f)

import os
evid_embed_list = os.listdir('./drive/My Drive/LAB/COMP90042 A3/embeddings/evid_embedding')

class PredictDataset():

    def __init__(self, data, maxlen):

        #Store the contents of the file in a pandas dataframe
        self.data = data

    def __len__(self):
        return len(self.data)

    def __getitem__(self, index):

        #Selecting the sentence and label at the specified index in the data frame
        evid_id = self.data[index][0]
        evid_embed_vector = self.data[index][1]

        return evid_id, evid_embed_vector

BATCH_SIZE = 1024

torch.cuda.empty_cache()
gc.collect()

import copy

top_5_dev_evid = {}
processed_dev_embed = []

with open('./drive/My Drive/LAB/COMP90042 A3/predictions/Retrieval_BERT_dev.pickle', 'rb') as f:
    top_5_dev_evid = pickle.load(f)
  
with open('./drive/My Drive/LAB/COMP90042 A3/embeddings/processed_dev_embed.json', 'r') as f:
    processed_dev_embed = json.load(f)

for evid_embed_dir in evid_embed_list:

  if evid_embed_dir in processed_dev_embed:
    print('pass', evid_embed_dir)
    continue
  
  print(evid_embed_dir)
  
  with open(f'./drive/My Drive/LAB/COMP90042 A3/embeddings/evid_embedding/{evid_embed_dir}', 'r') as f:
    evid_embed = json.load(f)
    evid_embeds = [(evid_id, evid_embed[evid_id]) for evid_id in evid_embed]

    for dev_id in dev_embed:

      dev_embed_vector = torch.Tensor([dev_embed[dev_id] for i in range(BATCH_SIZE)])
      dev_embed_vector = dev_embed_vector.cuda(gpu)
      
      if dev_id in top_5_dev_evid:
        tmp_evid = copy.deepcopy(top_5_dev_evid[dev_id])
      else:
        tmp_evid = []

      with torch.no_grad():
        ids = []
        evid_embed_vectors = []
        for i in range(len(evid_embeds)):
          ids.append(evid_embeds[i][0])
          evid_embed_vectors.append(evid_embeds[i][1])
        
          if (i+1) % BATCH_SIZE == 0:
            evid_embed_vectors = torch.Tensor(evid_embed_vectors)
            evid_embed_vectors = evid_embed_vectors.cuda(gpu)

            predict = net.neural_layer(dev_embed_vector, evid_embed_vectors, BATCH_SIZE).tolist()

            for j in range(BATCH_SIZE):
              tmp_evid.append((ids[j], predict[j][0]))

            ids = []
            evid_embed_vectors = []
        

      tmp_evid.sort(reverse=True, key = lambda x:x[1])
      top_5_dev_evid[dev_id] = tmp_evid[:5]

  with open('./drive/My Drive/LAB/COMP90042 A3/predictions/Retrieval_BERT_dev.pickle', 'wb') as f:
    pickle.dump(top_5_dev_evid, f)
  
  processed_dev_embed.append(evid_embed_dir)
  with open('./drive/My Drive/LAB/COMP90042 A3/embeddings/processed_dev_embed.json', 'w') as f:
    json.dump(processed_dev_embed, f)

BATCH_SIZE = 1024

torch.cuda.empty_cache()
gc.collect()

top_5_test_evid = {}
processed_test_embed = []

with open('./drive/My Drive/LAB/COMP90042 A3/predictions/Retrieval_BERT_test.pickle', 'rb') as f:
    top_5_test_evid = pickle.load(f)

with open('./drive/My Drive/LAB/COMP90042 A3/embeddings/processed_test_embed.json', 'r') as f:
    processed_test_embed = json.load(f)

for evid_embed_dir in evid_embed_list:

  if evid_embed_dir in processed_test_embed:
    print('pass', evid_embed_dir)
    continue
  
  print(evid_embed_dir)
  
  with open(f'./drive/My Drive/LAB/COMP90042 A3/embeddings/evid_embedding/{evid_embed_dir}', 'r') as f:
    evid_embed = json.load(f)
    evid_embeds = [(evid_id, evid_embed[evid_id]) for evid_id in evid_embed]

    for test_id in test_embed:

      test_embed_vector = torch.Tensor([test_embed[test_id] for i in range(BATCH_SIZE)])
      test_embed_vector = test_embed_vector.cuda(gpu)
      
      if test_id in top_5_test_evid:
        tmp_evid = copy.deepcopy(top_5_test_evid[test_id])
      else:
        tmp_evid = []

      with torch.no_grad():
        ids = []
        evid_embed_vectors = []
        for i in range(len(evid_embeds)):
          ids.append(evid_embeds[i][0])
          evid_embed_vectors.append(evid_embeds[i][1])
        
          if (i+1) % BATCH_SIZE == 0:
            evid_embed_vectors = torch.Tensor(evid_embed_vectors)
            evid_embed_vectors = evid_embed_vectors.cuda(gpu)

            predict = net.neural_layer(test_embed_vector, evid_embed_vectors, BATCH_SIZE).tolist()

            for j in range(BATCH_SIZE):
              tmp_evid.append((ids[j], predict[j][0]))

            ids = []
            evid_embed_vectors = []
        

      tmp_evid.sort(reverse=True, key = lambda x:x[1])
      top_5_test_evid[test_id] = tmp_evid[:5]

  with open('./drive/My Drive/LAB/COMP90042 A3/predictions/Retrieval_BERT_test.pickle', 'wb') as f:
    pickle.dump(top_5_test_evid, f)
  
  processed_test_embed.append(evid_embed_dir)
  with open('./drive/My Drive/LAB/COMP90042 A3/embeddings/processed_test_embed.json', 'w') as f:
    json.dump(processed_test_embed, f)

BATCH_SIZE = 1024

torch.cuda.empty_cache()
gc.collect()

top_5_future_evid = {}
processed_future_embed = []

for evid_embed_dir in evid_embed_list:

  if evid_embed_dir in processed_future_embed:
    print('pass', evid_embed_dir)
    continue
  
  print(evid_embed_dir)
  
  with open(f'./drive/My Drive/LAB/COMP90042 A3/embeddings/evid_embedding/{evid_embed_dir}', 'r') as f:
    evid_embed = json.load(f)
    evid_embeds = [(evid_id, evid_embed[evid_id]) for evid_id in evid_embed]

    for future_id in future_embed:

      future_embed_vector = torch.Tensor([future_embed[future_id] for i in range(BATCH_SIZE)])
      future_embed_vector = future_embed_vector.cuda(gpu)
      
      if future_id in top_5_future_evid:
        tmp_evid = copy.deepcopy(top_5_future_evid[future_id])
      else:
        tmp_evid = []

      with torch.no_grad():
        ids = []
        evid_embed_vectors = []
        for i in range(len(evid_embeds)):
          ids.append(evid_embeds[i][0])
          evid_embed_vectors.append(evid_embeds[i][1])
        
          if (i+1) % BATCH_SIZE == 0:
            evid_embed_vectors = torch.Tensor(evid_embed_vectors)
            evid_embed_vectors = evid_embed_vectors.cuda(gpu)

            predict = net.neural_layer(future_embed_vector, evid_embed_vectors, BATCH_SIZE).tolist()

            for j in range(BATCH_SIZE):
              tmp_evid.append((ids[j], predict[j][0]))

            ids = []
            evid_embed_vectors = []
        

      tmp_evid.sort(reverse=True, key = lambda x:x[1])
      top_5_future_evid[future_id] = tmp_evid[:5]

  with open('./drive/My Drive/LAB/COMP90042 A3/predictions/Retrieval_BERT_future.pickle', 'wb') as f:
    pickle.dump(top_5_future_evid, f)
  
  processed_future_embed.append(evid_embed_dir)
  with open('./drive/My Drive/LAB/COMP90042 A3/embeddings/processed_future_embed.pickle', 'wb') as f:
    pickle.dump(processed_future_embed, f)

F score

class PredictDataset():

    def __init__(self, data, maxlen):

        #Store the contents of the file in a pandas dataframe
        self.data = data

        #Initialize the BERT tokenizer
        self.tokenizer = BertTokenizer.from_pretrained('bert-base-uncased')

        self.maxlen = maxlen

    def __len__(self):
        return len(self.data)

    def __getitem__(self, index):

        #Selecting the sentence and label at the specified index in the data frame
        sentence1 = self.data[index][0][0]
        sentence2 = self.data[index][0][1]

        #Preprocessing the text to be suitable for BERT
        tokens1 = tokenizer.tokenize(sentence1)
        tokens2 = tokenizer.tokenize(sentence2)
        tokens1 = ['[CLS]'] + tokens1 + ['[SEP]']
        tokens2 = ['[CLS]'] + tokens2 + ['[SEP]']
        if len(tokens1) < self.maxlen:
            tokens1 = tokens1 + ['[PAD]' for _ in range(self.maxlen - len(tokens1))] #Padding sentences
        else:
            tokens1 = tokens1[:self.maxlen-1] + ['[SEP]'] #Prunning the list to be of specified max length

        if len(tokens2) < self.maxlen:
            tokens2 = tokens2 + ['[PAD]' for _ in range(self.maxlen - len(tokens2))] #Padding sentences
        else:
            tokens2 = tokens2[:self.maxlen-1] + ['[SEP]'] #Prunning the list to be of specified max length

        tokens_ids1 = self.tokenizer.convert_tokens_to_ids(tokens1) #Obtaining the indices of the tokens in the BERT Vocabulary
        tokens_ids_tensor1 = torch.tensor(tokens_ids1) #Converting the list to a pytorch tensor

        tokens_ids2 = self.tokenizer.convert_tokens_to_ids(tokens2) #Obtaining the indices of the tokens in the BERT Vocabulary
        tokens_ids_tensor2 = torch.tensor(tokens_ids2) #Converting the list to a pytorch tensor

        #Obtaining the attention mask i.e a tensor containing 1s for no padded tokens and 0s for padded ones
        attn_mask1 = (tokens_ids_tensor1 != 0).long()
        attn_mask2 = (tokens_ids_tensor2 != 0).long()



        return tokens_ids_tensor1, attn_mask1, tokens_ids_tensor2, attn_mask2

torch.cuda.empty_cache()
gc.collect()

EVIDENCE = [(claim,evidence[claim]) for claim in scientific_claims_id]

def get_retrievals(claims, file_name, SIZE=32):
  try:
    with open(f'./drive/My Drive/LAB/COMP90042 A3/predictions/Retrievals/{file_name}.pickle', 'rb') as f:
      retrievals = pickle.load(f)
  except:
      retrievals = {}

  EVIDENCE.sort(key = lambda x:x[0])

  i = 0
  for id in claims:
    if id in retrievals:
      print('pass:', id)
      continue

    print(id)
    data_for_predict = []

    claim_text = claims[id]['claim_text']

    for evid in EVIDENCE:

      evid_text = evid[1]

      data_for_predict.append((claim_text, evid_text))

    set_for_predict = PredictDataset(data_for_predict, maxlen = 512)

    predict_loader = DataLoader(set_for_predict, batch_size = SIZE, num_workers = 2)
    
    predicted_logit = list()
    with torch.no_grad():
      for it, (tokens_ids_tensor1, attn_mask1, tokens_ids_tensor2, attn_mask2) in enumerate(predict_loader):
        
        torch.cuda.empty_cache()
        gc.collect()
        tokens_ids_tensor1, attn_mask1, tokens_ids_tensor2, attn_mask2 = tokens_ids_tensor1.cuda(gpu), attn_mask1.cuda(gpu), tokens_ids_tensor2.cuda(gpu), attn_mask2.cuda(gpu)

        logits = net(tokens_ids_tensor1, attn_mask1, tokens_ids_tensor2, attn_mask2).tolist()
        logits = [x[0] for x in logits]
        
        for prediction in zip([claim_id for claim_id in EVIDENCE[it*SIZE:(it+1)*SIZE+1]], logits):
          predicted_logit.append(prediction)
      
      predicted_logit.sort(key = lambda x:x[1], reverse = True)
      predicted_logit = predicted_logit[:5]

      retrievals[id] = {'evidences': [x[0] for x in predicted_logit]}

      with open(f'./drive/My Drive/LAB/COMP90042 A3/predictions/Retrievals/{file_name}.pickle', 'wb') as f:
          pickle.dump(retrievals, f)

get_retrievals(dev_claims, 'Retrieval_SBERT_reduced_dev', SIZE=32)

get_retrievals(test_claims, 'Retrieval_SBERT_reduced_test', SIZE=32)

get_retrievals(future_claims, 'Retrieval_SBERT_reduced_future', SIZE=32)

# with open('./drive/My Drive/LAB/COMP90042 A3/notebooks/test_predictions_sentence.pickle', 'rb') as f:
#   final_test_predictions_sentence = pickle.load(f)

# final_test_predictions_sentence = {}

# SIZE = 32

# EVIDENCE.sort(key = lambda x:x[0])

# i = 0
# for id in test_claims:
#   if id in final_test_predictions_sentence:
#     print('pass:', id)
#     continue

#   print(id)
#   test_data_for_predict = []

#   claim_text = test_claims[id]['claim_text']

#   for evid in EVIDENCE:

#     evid_text = evid[1]

#     test_data_for_predict.append((claim_text, evid_text))

#   test_set_for_predict = PredictDataset(test_data_for_predict, maxlen = 512)

#   test_predict_loader = DataLoader(test_set_for_predict, batch_size = SIZE, num_workers = 2)
  
#   predicted_logit = list()
#   with torch.no_grad():
#     for it, (tokens_ids_tensor1, attn_mask1, tokens_ids_tensor2, attn_mask2) in enumerate(dev_predict_loader):
      
#       torch.cuda.empty_cache()
#       gc.collect()
#       tokens_ids_tensor1, attn_mask1, tokens_ids_tensor2, attn_mask2 = tokens_ids_tensor1.cuda(gpu), attn_mask1.cuda(gpu), tokens_ids_tensor2.cuda(gpu), attn_mask2.cuda(gpu)

#       logits = net(tokens_ids_tensor1, attn_mask1, tokens_ids_tensor2, attn_mask2).tolist()
#       logits = [x[0] for x in logits]
      
#       for prediction in zip([claim_id for claim_id in EVIDENCE[it*SIZE:(it+1)*SIZE+1]], logits):
#         predicted_logit.append(prediction)
    
#     predicted_logit.sort(key = lambda x:x[1], reverse = True)
#     predicted_logit = predicted_logit[:5]

#     final_test_predictions_sentence[id] = {'evidences': [x[0] for x in predicted_logit]}

#     with open('./drive/My Drive/LAB/COMP90042 A3/notebooks/test_predictions_sentence.pickle', 'wb') as f:
#         pickle.dump(final_test_predictions_sentence, f)
