==> notebooks/5_2_Classifier_Bert_2.ipynb <==
# 5.3 Classifier BERT

from google.colab import drive
drive.mount('/content/drive')

## Read in Data

import json
import numpy as np
import pickle
import gc

with open ('./drive/My Drive/LAB/COMP90042 A3/data/curated/train_claims2.json') as f:
    train_claims = json.load(f)

with open ('./drive/My Drive/LAB/COMP90042 A3/data/raw/dev-claims.json') as f:
    dev_claims = json.load(f)

with open ('./drive/My Drive/LAB/COMP90042 A3/data/curated/test_claims2.json') as f:
    test_claims = json.load(f)

with open ('./drive/My Drive/LAB/COMP90042 A3/data/raw/evidence.json') as f:
    evidence = json.load(f)

with open ('./drive/My Drive/LAB/COMP90042 A3/data/raw/test-claims-unlabelled.json') as f:
    future_claims = json.load(f)

import random
random.seed(19260817)

evid_id_list = [evid_id for evid_id in evidence]

### Create dataset used to train retriever

ENCODING = {'REFUTES': 0, 'DISPUTED': 1, 'NOT_ENOUGH_INFO': 2, 'SUPPORTS': 3}
DECODING = {ENCODING[key]:key for key in ENCODING}

training_data = []

for id in train_claims:

  claim_text = train_claims[id]['claim_text']

  label = ENCODING[train_claims[id]['claim_label']] # LABELS MUST BE ENCODED!!!!!!

  n_evid = len(train_claims[id]['evidences'])
  
  for evid_id in train_claims[id]['evidences']:
    evid_text = evidence[evid_id]

    training_data.append(((claim_text, evid_text), label))

dev_data = []

for id in dev_claims:

  claim_text = dev_claims[id]['claim_text']

  label = ENCODING[dev_claims[id]['claim_label']] # LABELS MUST BE ENCODED!!!!!!

  n_evid = len(dev_claims[id]['evidences'])
  
  for evid_id in dev_claims[id]['evidences']:
    evid_text = evidence[evid_id]
    
    dev_data.append(((claim_text, evid_text), label))

test_data = []

for id in test_claims:

  claim_text = test_claims[id]['claim_text']

  label = ENCODING[test_claims[id]['claim_label']] # LABELS MUST BE ENCODED!!!!!!

  n_evid = len(test_claims[id]['evidences'])
  
  for evid_id in test_claims[id]['evidences']:
    evid_text = evidence[evid_id]
    
    test_data.append(((claim_text, evid_text), label))

scientific_claims_id = set()
for claim in train_claims:
  for evid in train_claims[claim]['evidences']:
    scientific_claims_id.add(evid)

for claim in dev_claims:
  for evid in dev_claims[claim]['evidences']:
    scientific_claims_id.add(evid)

for claim in test_claims:
  for evid in test_claims[claim]['evidences']:
    scientific_claims_id.add(evid)


scientific_claims_id = list(scientific_claims_id)

!pip install torch torchvision transformers

## Build model, dataloader etc

from transformers import BertModel

from transformers import BertTokenizer

tokenizer = BertTokenizer.from_pretrained('bert-base-uncased')

import torch
from torch.utils.data import Dataset
from transformers import BertTokenizer
import pandas as pd

class Dataset():

    def __init__(self, data, maxlen):

        #Store the contents of the file in a pandas dataframe
        self.data = data

        #Initialize the BERT tokenizer
        self.tokenizer = BertTokenizer.from_pretrained('bert-base-uncased')

        self.maxlen = maxlen

    def __len__(self):
        return len(self.data)

    def __getitem__(self, index):

        #Selecting the sentence and label at the specified index in the data frame
        sentence1 = self.data[index][0][0]
        sentence2 = self.data[index][0][1]
        label = self.data[index][1]

        #Preprocessing the text to be suitable for BERT
        tokens1 = tokenizer.tokenize(sentence1)
        tokens2 = tokenizer.tokenize(sentence2)
        tokens1 = ['[CLS]'] + tokens1 + ['[SEP]']
        tokens2 = tokens2 + ['[SEP]']
        tokens = tokens1 + tokens2 #Insering the CLS and SEP token in the beginning and end of the sentence
        if len(tokens) < self.maxlen:
            tokens = tokens + ['[PAD]' for _ in range(self.maxlen - len(tokens))] #Padding sentences
        else:
            tokens = tokens[:self.maxlen-1] + ['[SEP]'] #Prunning the list to be of specified max length

        tokens_ids = self.tokenizer.convert_tokens_to_ids(tokens) #Obtaining the indices of the tokens in the BERT Vocabulary
        tokens_ids_tensor = torch.tensor(tokens_ids) #Converting the list to a pytorch tensor

        #Obtaining the attention mask i.e a tensor containing 1s for no padded tokens and 0s for padded ones
        attn_mask = (tokens_ids_tensor != 0).long()

        seg_ids = [0 for _ in range(len(tokens1))]
        seg_ids2 = [1 for _ in range(self.maxlen-len(tokens1))]
        seg_ids.extend(seg_ids2)

        seg_ids_tensor = torch.tensor(seg_ids)


        return tokens_ids_tensor, attn_mask, seg_ids_tensor, label

from torch.utils.data import DataLoader

#Creating instances of training and development set
#maxlen sets the maximum length a sentence can have
#any sentence longer than this length is truncated to the maxlen size
train_set = Dataset(training_data, maxlen = 512)
dev_set = Dataset(dev_data, maxlen = 512)
test_set = Dataset(test_data, maxlen = 512)

#Creating intsances of training and development dataloaders
train_loader = DataLoader(train_set, batch_size = 16, shuffle = True, num_workers = 2)
dev_loader = DataLoader(dev_set, batch_size = 16, shuffle = True, num_workers = 2)
test_loader = DataLoader(test_set, batch_size = 16, shuffle = True, num_workers = 2)

import torch
import torch.nn as nn
from transformers import BertModel

class nEvidClassifier(nn.Module):

    def __init__(self):
        super(nEvidClassifier, self).__init__()
        #Instantiating BERT model object 
        self.bert_layer = BertModel.from_pretrained('bert-base-uncased')
        
        #Classification layer
        #input dimension is 768 because [CLS] embedding has a dimension of 768
        #output dimension is 1 because we're working with a binary classification problem
        self.cls_layer = nn.Linear(768, 30)

        self.cls_layer2 = nn.Linear(30, 4)

        self.softmax_layer = nn.Softmax(dim=4)

    def forward(self, seq, attn_masks, seg_ids):
        '''
        Inputs:
            -seq : Tensor of shape [B, T] containing token ids of sequences
            -attn_masks : Tensor of shape [B, T] containing attention masks to be used to avoid contibution of PAD tokens
        '''

        #Feeding the input to BERT model to obtain contextualized representations
        outputs = self.bert_layer(seq, attention_mask = attn_masks, token_type_ids = seg_ids, return_dict=True)
        cont_reps = outputs.last_hidden_state

        #Obtaining the representation of [CLS] head (the first token)
        cls_rep = cont_reps[:, 0]

        #Feeding cls_rep to the classifier layer
        output = self.cls_layer(cls_rep)

        output = self.cls_layer2(output)

        return output

gpu = 0 #gpu ID

print("Creating the sentiment regressor, initialised with pretrained BERT-BASE parameters...")
net = nEvidClassifier()
net.cuda(gpu) #Enable gpu support for the model
print("Done creating the sentiment regressor.")

### Setup Training

import torch.nn as nn
import torch.optim as optim

criterion = nn.CrossEntropyLoss()
opti = optim.Adam(net.parameters(), lr = 2e-5)

def get_accuracy_from_logits(pseudo_probs, labels):

    correct = 0
    total = 0
    for i in range(len(labels)):
        _, predicted = torch.max(pseudo_probs.data, 1)
        total += labels.size(0)
        correct += (predicted == labels).sum().item()

    return correct/total

    # index = torch.argmax(pseudo_probs.unsqueeze(-1))
    # soft_probs = (probs > 0.5).long()
    # acc = (soft_probs.squeeze() == labels).float().mean()
    # return acc
    

def evaluate(net, criterion, dataloader, gpu):
    net.eval()

    mean_acc, mean_loss = 0, 0
    count = 0
    acc = 0

    with torch.no_grad():
        for seq, attn_masks, seg_ids_tensor, labels in dataloader:
            seq, attn_masks, seg_ids_tensor, labels = seq.cuda(gpu), attn_masks.cuda(gpu), seg_ids_tensor.cuda(gpu), labels.cuda(gpu)
            output = net(seq, attn_masks, seg_ids_tensor)
            _, predicted = torch.max(output.data, 1)
            acc += (predicted == labels).sum().item() / len(labels)
            mean_loss += criterion(output, labels).item()
            count += 1

    return acc / count, mean_loss / count

import time

def train(net, criterion, opti, train_loader, dev_loader, max_eps, gpu):

    best_acc = 0
    st = time.time()
    for ep in range(max_eps):
        
        net.train()
        for it, (seq, attn_masks, seg_ids_tensor, labels) in enumerate(train_loader):
            #Clear gradients
            opti.zero_grad()  
            #Converting these to cuda tensors
            seq, attn_masks, seg_ids_tensor, labels = seq.cuda(gpu), attn_masks.cuda(gpu), seg_ids_tensor.cuda(gpu), labels.cuda(gpu)

            #Obtaining the logits from the model
            output = net(seq, attn_masks, seg_ids_tensor)

            #Computing loss
            loss = criterion(output, labels)

            #Backpropagating the gradients
            loss.backward()

            #Optimization step
            opti.step()
              
            if it % 100 == 0:
                
                _, predicted = torch.max(output.data, 1)
                acc = (predicted == labels).sum().item() / len(labels)
                print("Iteration {} of epoch {} complete. Loss: {}; Accuracy: {}; Time taken (s): {}".format(it, ep, loss.item(), acc, (time.time()-st)))
                st = time.time()

        
        dev_acc, dev_loss = evaluate(net, criterion, dev_loader, gpu)
        print("Epoch {} complete! Development Accuracy: {}; Development Loss: {}".format(ep, dev_acc, dev_loss))
        if dev_acc > best_acc:
            print("Best development accuracy improved from {} to {}, saving model...".format(best_acc, dev_acc))
            best_acc = dev_acc
            torch.save(net.state_dict(), './drive/My Drive/LAB/COMP90042 A3/models/Classifiers/Classifier_Bert_2.dat')
            torch.save(net, './drive/My Drive/LAB/COMP90042 A3/models/Classifiers/Classifier_Bert_2.pt')

torch.cuda.empty_cache() 

num_epoch = 5

#fine-tune the model
train(net, criterion, opti, train_loader, dev_loader, num_epoch, gpu)

## Read In Model

net.load_state_dict(torch.load('./drive/My Drive/LAB/COMP90042 A3/models/Classifiers/Classifier_Bert_2.dat'))

net = torch.load('./drive/My Drive/LAB/COMP90042 A3/models/Classifiers/Classifier_Bert_2.pt')

net.eval()

## Make Predictions (make Classifications)

class PredictDataset():

    def __init__(self, data, maxlen):

        #Store the contents of the file in a pandas dataframe
        self.data = data

        #Initialize the BERT tokenizer
        self.tokenizer = BertTokenizer.from_pretrained('bert-base-uncased')

        self.maxlen = maxlen

    def __len__(self):
        return len(self.data)

    def __getitem__(self, index):

        #Selecting the sentence and label at the specified index in the data frame
        sentence1 = self.data[index][0]
        sentence2 = self.data[index][1]

        #Preprocessing the text to be suitable for BERT
        tokens1 = tokenizer.tokenize(sentence1)
        tokens2 = tokenizer.tokenize(sentence2)
        tokens1 = ['[CLS]'] + tokens1 + ['[SEP]']
        tokens2 = tokens2 + ['[SEP]']
        tokens = tokens1 + tokens2 #Insering the CLS and SEP token in the beginning and end of the sentence
        if len(tokens) < self.maxlen:
            tokens = tokens + ['[PAD]' for _ in range(self.maxlen - len(tokens))] #Padding sentences
        else:
            tokens = tokens[:self.maxlen-1] + ['[SEP]'] #Prunning the list to be of specified max length

        tokens_ids = self.tokenizer.convert_tokens_to_ids(tokens) #Obtaining the indices of the tokens in the BERT Vocabulary
        tokens_ids_tensor = torch.tensor(tokens_ids) #Converting the list to a pytorch tensor

        #Obtaining the attention mask i.e a tensor containing 1s for no padded tokens and 0s for padded ones
        attn_mask = (tokens_ids_tensor != 0).long()

        seg_ids = [0 for _ in range(len(tokens1))]
        seg_ids2 = [1 for _ in range(self.maxlen-len(tokens1))]
        seg_ids.extend(seg_ids2)

        seg_ids_tensor = torch.tensor(seg_ids)


        return tokens_ids_tensor, attn_mask, seg_ids_tensor

def get_classif(claims, file_name, SIZE):


  try:
    with open(f'./drive/My Drive/LAB/COMP90042 A3/predictions/Retrievals/{file_name}.pickle', 'rb') as f:
      predictions = pickle.load(f)
  except:
      predictions = {}


  i = 0
  for id in claims:
    if id in predictions:
      print('pass:', id)
      continue

    print(id)

    claim_text = claims[id]['claim_text']
    
    data_for_predict = list()

    for evid_id in claims[id]['evidences']:
      evid_text = evidence[evid_id]
      
      data_for_predict.append((claim_text, evid_text))

    set_for_predict = PredictDataset(data_for_predict, maxlen = 512)

    predict_loader = DataLoader(set_for_predict, batch_size = SIZE, num_workers = 2)
    
    predicted_logit = list()
    with torch.no_grad():
      for it, (seq, attn_masks, seg_ids_tensor) in enumerate(predict_loader):
        
        torch.cuda.empty_cache()
        gc.collect()
        seq, attn_masks, seg_ids_tensor = seq.cuda(gpu), attn_masks.cuda(gpu), seg_ids_tensor.cuda(gpu)

        logits = net(seq, attn_masks, seg_ids_tensor).tolist()
      
      predictions[id] = logits

      with open(f'./drive/My Drive/LAB/COMP90042 A3/predictions/Classifications/{file_name}.pickle', 'wb') as f:
          pickle.dump(predictions, f)

def get_classif_from_pred_retrievals(claims, predicted_evidence, file_name, SIZE):


  try:
    with open(f'./drive/My Drive/LAB/COMP90042 A3/predictions/Retrievals/{file_name}.pickle', 'rb') as f:
      predictions = pickle.load(f)
  except:
      predictions = {}


  i = 0
  for id in claims:
    if id in predictions:
      print('pass:', id)
      continue

    print(id)

    claim_text = claims[id]['claim_text']
    
    data_for_predict = list()

    for evid_id in predicted_evidence[id]:
      evid_text = evidence[evid_id]
      
      data_for_predict.append((claim_text, evid_text))

    set_for_predict = PredictDataset(data_for_predict, maxlen = 512)

    predict_loader = DataLoader(set_for_predict, batch_size = SIZE, num_workers = 2)
    
    predicted_logit = list()
    with torch.no_grad():
      for it, (seq, attn_masks, seg_ids_tensor) in enumerate(predict_loader):
        
        torch.cuda.empty_cache()
        gc.collect()
        seq, attn_masks, seg_ids_tensor = seq.cuda(gpu), attn_masks.cuda(gpu), seg_ids_tensor.cuda(gpu)

        logits = net(seq, attn_masks, seg_ids_tensor).tolist()
      
      predictions[id] = logits

      with open(f'./drive/My Drive/LAB/COMP90042 A3/predictions/Classifications/{file_name}.pickle', 'wb') as f:
          pickle.dump(predictions, f)

CLASS_MODEL_A = 'BERT2'

Classifications from Ground Truth

get_classif(dev_claims, f'Classification_{CLASS_MODEL_A}_dev', SIZE=32)

get_classif(test_claims, f'Classification_{CLASS_MODEL_A}_test', SIZE=32)

CLASS_MODEL = "BERT2"
OPEN_PREDICTION = "SBERTspecial"
RETRIEV_MODEL = "sberts"
K = 5

with open(f'./drive/My Drive/LAB/COMP90042 A3/predictions/Retrievals/Retrieval_{OPEN_PREDICTION}_dev.pickle', 'rb') as f:
    dev_predicted_evidence = pickle.load(f)

dev_claims_pred = {}

for claim in dev_predicted_evidence:
  # evid_list = [x[0] for x in dev_predicted_evidence[claim]['evidences']]
  evid_list = [x[0] for x in dev_predicted_evidence[claim]]

  dev_claims_pred[claim] = evid_list[:K]

with open(f'./drive/My Drive/LAB/COMP90042 A3/predictions/Retrievals/Retrieval_{OPEN_PREDICTION}_test.pickle', 'rb') as f:
    test_predicted_evidence = pickle.load(f)

test_claims_pred = {}

for claim in test_predicted_evidence:
  # evid_list = [x[0] for x in test_predicted_evidence[claim]['evidences']]
  evid_list = [x[0] for x in test_predicted_evidence[claim]]

  test_claims_pred[claim] = evid_list[:K]

with open(f'./drive/My Drive/LAB/COMP90042 A3/predictions/Retrievals/Retrieval_{OPEN_PREDICTION}_future.pickle', 'rb') as f:
    future_predicted_evidence = pickle.load(f)

future_claims_pred = {}

for claim in future_predicted_evidence:
  # evid_list = [x[0] for x in future_predicted_evidence[claim]['evidences']]
  evid_list = [x[0] for x in future_predicted_evidence[claim]]

  future_claims_pred[claim] = evid_list[:K]

Classifications from Predicted Evidence

get_classif_from_pred_retrievals(dev_claims, dev_claims_pred, f'Classification_{CLASS_MODEL}_{RETRIEV_MODEL}{K}_devp', SIZE=32)

get_classif_from_pred_retrievals(test_claims, test_claims_pred, f'Classification_{CLASS_MODEL}_{RETRIEV_MODEL}{K}_testp', SIZE=32)

get_classif_from_pred_retrievals(future_claims, future_claims_pred, f'Classification_{CLASS_MODEL}_{RETRIEV_MODEL}{K}_futurep', SIZE=32)
